requirements.txt
numpy
scipy
tensorflow-gpu
python_speech_features
pandas
matplotlib
pyxdg
